[Release] merge branch 'development' into 'implementation'

+++ back-end/hospital-api/src/main/java/net/pladema/staff/controller/mapper/BackofficeEpisodeDocumentTypeStore.java
@@ -3,13 +3,13 @@ package net.pladema.staff.controller.mapper;
 import net.pladema.establishment.repository.EpisodeDocumentTypeRepository;
 import net.pladema.sgx.backoffice.repository.BackofficeStore;
 
+import net.pladema.sgx.exceptions.BackofficeValidationException;
 import net.pladema.staff.repository.entity.EpisodeDocumentType;
 
 import org.springframework.data.domain.Example;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -30,7 +30,7 @@ public class BackofficeEpisodeDocumentTypeStore implements BackofficeStore<Episo
 				PageRequest.of(
 						pageable.getPageNumber(),
 						pageable.getPageSize(),
-						Sort.by(Sort.Direction.ASC, "description")
+						pageable.getSort()
 				)
 		);
 	}
@@ -52,12 +52,26 @@ public class BackofficeEpisodeDocumentTypeStore implements BackofficeStore<Episo
 
 	@Override
 	public EpisodeDocumentType save(EpisodeDocumentType entity) {
+		if (entity.getConsentId() == null || entity.getConsentId() == 1) {
+			entity.setConsentId(1);
+			entity.setRichTextBody(null);
+		}
+
+		if (entity.getConsentId() != 1
+			&& entity.getId() == null
+			&& repository.existsConsentDocumentById(entity.getConsentId()))
+			throw new BackofficeValidationException("Ya existe ese documento de consentimiento");
 		return repository.save(entity);
 	}
 
 	@Override
 	public void deleteById(Integer id) {
-		repository.deleteById(id);
+		repository.findById(id).ifPresent(episodeDocumentType -> {
+			if (episodeDocumentType.getConsentId().equals(1))
+				repository.deleteById(id);
+			else
+				throw new BackofficeValidationException("El documento es de consentimiento y no se puede eliminar");
+		});
 	}
 
 	@Override
